webapp/requirements.txt
django==4.2.3
psycopg2==2.9.6
uWSGI==2.0.21



celery==5.3.1
django-celery-beat==2.5.0
django-cors-headers==4.1.0
djangorestframework==3.14.0
redis==4.6.0
stripe==5.4.0
pillow==10.0.0
google-auth==2.21.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
requests==2.31.0
pyjwt==2.7.0
drf-yasg==1.21.6
django-jazzmin==2.6.0
